5.4.2 exercise
Print only game-related info to the console.
• Log invalid guesses. Pick an appropriate logging level.
• Log invalid words read from file data.txt. Log them at a severe level. This is because data.txt is
under our control, so invalid input could be severe, and should be cleaned up.
• Log all valid words from file data.txt. Don’t print them.
• Log all exceptions. You can use logger.log(Level.WARNING, "Your message.", e) to log exception
e. Pick a level appropriate to the seriousness of the exception. Do not print traces to screen. If
you want to inform the user that something went wrong, use a message with generic wording.
• Objects of class SQLiteConnectionManager should not print anything to console. Remember to log all exceptions